02-01-2022_demo

+++ application/controllers/Poliklinik.php
@@ -163,6 +163,7 @@ class Poliklinik extends CI_Controller {
 			$row[]   = $unit->nadokter;
 			$row[]   = $jpas;
 			$row[]   = $unit->nobpjs;
+			$row[]	 = ($unit->nobpjs == "" ? "" : "<a href='". base_url("pcare/". $unit->noreg) ."' target='_blank'><button type='button' class='btn btn-primary btn-xs'>Bridging</button></a>");
 						//saya mengganti line 120 semula $unit->id
 
 
@@ -394,8 +395,8 @@ class Poliklinik extends CI_Controller {
 					'tarifrs'     	=> $datatarif->tarifrspoli,
 					'tarifdr'     	=> $datatarif->tarifdrpoli,
 					'paramedis'		=> $datatarif->feemedispoli,
-					'obatpoli'		=> $datatarif->bhppoli,
-					// 'obatpoli'		=> $datatarif->obatpoli,
+					// 'obatpoli'		=> $datatarif->bhppoli,
+					'obatpoli'		=> $datatarif->obatpoli,
 					'bahan'       	=> 0,
 					'koperawat'   	=> $_perawat,
 				);
@@ -434,42 +435,44 @@ class Poliklinik extends CI_Controller {
 			$c_hrgalkes  = $this->input->post('hrgalkes');
 			$c_totalkes  = $this->input->post('totalkes');
 
-			$this->db->query("DELETE from tbl_alkestransaksi WHERE notr = '$noreg_dok' and koders='$cabang'");
+			if(isset($c_kdalkes)){
+				$this->db->query("DELETE from tbl_alkestransaksi WHERE notr = '$noreg_dok' and koders='$cabang'");
 
-			$jumdataalkes = count($c_kdalkes);
-			for($i=0;$i<=$jumdataalkes-1;$i++){
-				$_c_hrgalkes   = str_replace(',','',$c_hrgalkes[$i]);
-				$_c_totalkes   = str_replace(',','',$c_totalkes[$i]);
+				$jumdataalkes = count($c_kdalkes);
+				for($i=0;$i<=$jumdataalkes-1;$i++){
+					$_c_hrgalkes   = str_replace(',','',$c_hrgalkes[$i]);
+					$_c_totalkes   = str_replace(',','',$c_totalkes[$i]);
 
-				$data_alkes = array(
-					'koders'       => $cabang,
-					'notr'         => $noreg_dok,
-					'kodeobat'     => $c_kdalkes[$i],
-					'qty'          => $c_qtyalkes[$i],
-					'satuan'       => $c_satalkes[$i],
-					'harga'        => $_c_hrgalkes,
-					'totalharga'   => $_c_totalkes,
-					'tgltransaksi' => $tanggal,
-					'gudang'	   => $gudang_bhp,
-					'dibebankan'   => $c_bbn[$i],
-				);
+					$data_alkes = array(
+						'koders'       => $cabang,
+						'notr'         => $noreg_dok,
+						'kodeobat'     => $c_kdalkes[$i],
+						'qty'          => $c_qtyalkes[$i],
+						'satuan'       => $c_satalkes[$i],
+						'harga'        => $_c_hrgalkes,
+						'totalharga'   => $_c_totalkes,
+						'tgltransaksi' => $tanggal,
+						'gudang'	   => $gudang_bhp,
+						'dibebankan'   => $c_bbn[$i],
+					);
 
-				if($c_kdalkes[$i]!=""){
-					// Lose Stock
-					$check_stock	= $this->db->query("SELECT * FROM tbl_barangstock WHERE kodebarang = '". $c_kdalkes[$i] ."' AND koders = '$cabang' AND gudang = '$gudang_bhp'");
+					if($c_kdalkes[$i]!=""){
+						// Lose Stock
+						$check_stock	= $this->db->query("SELECT * FROM tbl_barangstock WHERE kodebarang = '". $c_kdalkes[$i] ."' AND koders = '$cabang' AND gudang = '$gudang_bhp'");
 
-					if($check_stock->num_rows() == 0){
-						$this->db->query("INSERT INTO tbl_barangstock (koders,kodebarang,gudang,keluar,saldoakhir) 
-						VALUES ('$cabang', '". $c_kdalkes[$i] ."', '$gudang_bhp', '". $c_qtyalkes[$i] ."', '". $c_qtyalkes[$i] ."')");
-					} else {
-						$this->db->query("UPDATE tbl_barangstock SET keluar = keluar+". $c_qtyalkes[$i] .", saldoakhir = saldoakhir-". $c_qtyalkes[$i] ." WHERE kodebarang = '". $c_kdalkes[$i] ."' AND koders = '$cabang'  AND gudang = '$gudang_bhp'");
+						if($check_stock->num_rows() == 0){
+							$this->db->query("INSERT INTO tbl_barangstock (koders,kodebarang,gudang,keluar,saldoakhir) 
+							VALUES ('$cabang', '". $c_kdalkes[$i] ."', '$gudang_bhp', '". $c_qtyalkes[$i] ."', '". $c_qtyalkes[$i] ."')");
+						} else {
+							$this->db->query("UPDATE tbl_barangstock SET keluar = keluar+". $c_qtyalkes[$i] .", saldoakhir = saldoakhir-". $c_qtyalkes[$i] ." WHERE kodebarang = '". $c_kdalkes[$i] ."' AND koders = '$cabang'  AND gudang = '$gudang_bhp'");
+						}
+
+						// $this->db->query("UPDATE tbl_barangstock SET keluar = keluar+". $c_qtyalkes[$i] .", saldoakhir = saldoakhir-". $c_qtyalkes[$i] ." WHERE kodebarang = '". $c_kdalkes[$i] ."' AND koders = '$cabang'  AND gudang = '$gudang_bhp'");
+
+						$insert_detil = $this->db->insert('tbl_alkestransaksi',$data_alkes);
 					}
 
-					// $this->db->query("UPDATE tbl_barangstock SET keluar = keluar+". $c_qtyalkes[$i] .", saldoakhir = saldoakhir-". $c_qtyalkes[$i] ." WHERE kodebarang = '". $c_kdalkes[$i] ."' AND koders = '$cabang'  AND gudang = '$gudang_bhp'");
-
-					$insert_detil = $this->db->insert('tbl_alkestransaksi',$data_alkes);
 				}
-
 			}
 
 			// -- ERESEP -- //
@@ -491,36 +494,38 @@ class Poliklinik extends CI_Controller {
 				urut_transaksi("ERESEP", 20);
 			}
 
-			$this->db->query("DELETE from tbl_eresep WHERE noreg = '$noreg_dok' and koders='$cabang'");
-			$jumdataer			= count($er_obat);
-			for($er=0;$er<=$jumdataer-1;$er++){
-				$nama_obat		= data_master("tbl_barang", array("kodebarang" => $er_obat[$er]))->namabarang;
-				$totalharga		= str_replace(",", "", $er_totalharga[$er]);
+			if(isset($er_obat)){
+				$this->db->query("DELETE from tbl_eresep WHERE noreg = '$noreg_dok' and koders='$cabang'");
+				$jumdataer			= count($er_obat);
+				for($er=0;$er<=$jumdataer-1;$er++){
+					$nama_obat		= data_master("tbl_barang", array("kodebarang" => $er_obat[$er]))->namabarang;
+					$totalharga		= str_replace(",", "", $er_totalharga[$er]);
 
-				$data_eresep_add	= array(
-					"koders"		=> $cabang,
-					"noreg"			=> $noreg_dok,
-					"orderno"		=> $orderno,
-					"kodeobat"		=> $er_obat[$er],
-					"namaobat"		=> $nama_obat,
-					"satuan"		=> $er_satuan[$er],
-					"qty"			=> $er_jmlobat[$er],
-					"qty_perhari"	=> $er_jmlhobathari[$er],
-					"jml_hari"		=> $er_jmlhari[$er],
-					"qty_minum"		=> $er_qtymin[$er],
-					"harga"			=> $er_harga[$er],
-					"totalharga"	=> $totalharga,
-					"aturanpakai"	=> $er_aturan[$er],
-					"gudang"		=> $gudang,
-					"resepno"		=> $orderno,
-					"namauser"		=> $userid,
-					"qtyawal"		=> $er_jmlobat[$er],
-					"keterangan"	=> $er_keterangan[$er],
-					"kronis"		=> $er_kronishide[$er],
-					// "validfarmasi"	=> $er_validfar[$er],
-				);
+					$data_eresep_add	= array(
+						"koders"		=> $cabang,
+						"noreg"			=> $noreg_dok,
+						"orderno"		=> $orderno,
+						"kodeobat"		=> $er_obat[$er],
+						"namaobat"		=> $nama_obat,
+						"satuan"		=> $er_satuan[$er],
+						"qty"			=> $er_jmlobat[$er],
+						"qty_perhari"	=> $er_jmlhobathari[$er],
+						"jml_hari"		=> $er_jmlhari[$er],
+						"qty_minum"		=> $er_qtymin[$er],
+						"harga"			=> $er_harga[$er],
+						"totalharga"	=> $totalharga,
+						"aturanpakai"	=> $er_aturan[$er],
+						"gudang"		=> $gudang,
+						"resepno"		=> $orderno,
+						"namauser"		=> $userid,
+						"qtyawal"		=> $er_jmlobat[$er],
+						"keterangan"	=> $er_keterangan[$er],
+						"kronis"		=> $er_kronishide[$er],
+						// "validfarmasi"	=> $er_validfar[$er],
+					);
 
-				$this->db->insert("tbl_eresep", $data_eresep_add);
+					$this->db->insert("tbl_eresep", $data_eresep_add);
+				}
 			}
 
 			// -- RACIKAN -- //
@@ -550,79 +555,85 @@ class Poliklinik extends CI_Controller {
 			$this->db->query("DELETE FROM tbl_eracik WHERE racikid = 'RACIK2' AND noreg = '$noreg_dok' AND koders = '$cabang'");
 			$this->db->query("DELETE FROM tbl_eracik WHERE racikid = 'RACIK3' AND noreg = '$noreg_dok' AND koders = '$cabang'");
 
-			$jumracik1	= count($ra_kodeobat1);
-			for($ra1 = 0; $ra1 <= $jumracik1-1; $ra1++){
-				$nama_obat_ra1	= data_master("tbl_barang", array("kodebarang" => $ra_kodeobat1[$ra1]))->namabarang;
-				$harga_ra1		= str_replace(",", "", $ra_harga1[$ra1]);
-				$totalharga_ra1 = str_replace(",", "", $ra_totalharga1[$ra1]);
+			if(isset($ra_kodeobat1)){
+				$jumracik1	= count($ra_kodeobat1);
+				for($ra1 = 0; $ra1 <= $jumracik1-1; $ra1++){
+					$nama_obat_ra1	= data_master("tbl_barang", array("kodebarang" => $ra_kodeobat1[$ra1]))->namabarang;
+					$harga_ra1		= str_replace(",", "", $ra_harga1[$ra1]);
+					$totalharga_ra1 = str_replace(",", "", $ra_totalharga1[$ra1]);
 
-				$data_racik1	= array(
-					"koders"		=> $cabang,
-					"orderno"		=> $orderno,
-					"racikid"		=> "RACIK1",
-					"noreg"			=> $noreg_dok,
-					"kodeobat"		=> $ra_kodeobat1[$ra1],
-					"namaobat"		=> $nama_obat_ra1,
-					"satuan"		=> $ra_satuan1[$ra1],
-					"qty_racik"		=> $ra_qtyracik1[$ra1],
-					"qty_jual"		=> $ra_qtyjual1[$ra1],
-					"harga"			=> $harga_ra1,
-					"totalharga"	=> $totalharga_ra1,
-					"aturanpakai"	=> $ra_aturanpakai1,
-					"proses"		=> 1,
-				);
+					$data_racik1	= array(
+						"koders"		=> $cabang,
+						"orderno"		=> $orderno,
+						"racikid"		=> "RACIK1",
+						"noreg"			=> $noreg_dok,
+						"kodeobat"		=> $ra_kodeobat1[$ra1],
+						"namaobat"		=> $nama_obat_ra1,
+						"satuan"		=> $ra_satuan1[$ra1],
+						"qty_racik"		=> $ra_qtyracik1[$ra1],
+						"qty_jual"		=> $ra_qtyjual1[$ra1],
+						"harga"			=> $harga_ra1,
+						"totalharga"	=> $totalharga_ra1,
+						"aturanpakai"	=> $ra_aturanpakai1,
+						"proses"		=> 1,
+					);
 
-				$this->db->insert("tbl_eracik", $data_racik1);
+					$this->db->insert("tbl_eracik", $data_racik1);
+				}
 			}
 
-			$jumracik2	= count($ra_kodeobat2);
-			for($ra2 = 0; $ra2 <= $jumracik2-1; $ra2++){
-				$nama_obat_ra2	= data_master("tbl_barang", array("kodebarang" => $ra_kodeobat2[$ra2]))->namabarang;
-				$harga_ra2		= str_replace(",", "", $ra_harga2[$ra2]);
-				$totalharga_ra2 = str_replace(",", "", $ra_totalharga2[$ra2]);
+			if(isset($ra_kodeobat2)){
+				$jumracik2	= count($ra_kodeobat2);
+				for($ra2 = 0; $ra2 <= $jumracik2-1; $ra2++){
+					$nama_obat_ra2	= data_master("tbl_barang", array("kodebarang" => $ra_kodeobat2[$ra2]))->namabarang;
+					$harga_ra2		= str_replace(",", "", $ra_harga2[$ra2]);
+					$totalharga_ra2 = str_replace(",", "", $ra_totalharga2[$ra2]);
 
-				$data_racik2	= array(
-					"koders"		=> $cabang,
-					"orderno"		=> $orderno,
-					"racikid"		=> "RACIK2",
-					"noreg"			=> $noreg_dok,
-					"kodeobat"		=> $ra_kodeobat2[$ra2],
-					"namaobat"		=> $nama_obat_ra2,
-					"satuan"		=> $ra_satuan2[$ra2],
-					"qty_racik"		=> $ra_qtyracik2[$ra2],
-					"qty_jual"		=> $ra_qtyjual2[$ra2],
-					"harga"			=> $harga_ra2,
-					"totalharga"	=> $totalharga_ra2,
-					"aturanpakai"	=> $ra_aturanpakai2,
-					"proses"		=> 1,
-				);
+					$data_racik2	= array(
+						"koders"		=> $cabang,
+						"orderno"		=> $orderno,
+						"racikid"		=> "RACIK2",
+						"noreg"			=> $noreg_dok,
+						"kodeobat"		=> $ra_kodeobat2[$ra2],
+						"namaobat"		=> $nama_obat_ra2,
+						"satuan"		=> $ra_satuan2[$ra2],
+						"qty_racik"		=> $ra_qtyracik2[$ra2],
+						"qty_jual"		=> $ra_qtyjual2[$ra2],
+						"harga"			=> $harga_ra2,
+						"totalharga"	=> $totalharga_ra2,
+						"aturanpakai"	=> $ra_aturanpakai2,
+						"proses"		=> 1,
+					);
 
-				$query_insert_racik = $this->db->insert("tbl_eracik", $data_racik2);
+					$query_insert_racik = $this->db->insert("tbl_eracik", $data_racik2);
+				}
 			}
 
-			$jumracik3	= count($ra_kodeobat3);
-			for($ra3 = 0; $ra3 <= $jumracik3-1; $ra3++){
-				$nama_obat_ra3	= data_master("tbl_barang", array("kodebarang" => $ra_kodeobat3[$ra3]))->namabarang;
-				$harga_ra3		= str_replace(",", "", $ra_harga3[$ra3]);
-				$totalharga_ra3 = str_replace(",", "", $ra_totalharga3[$ra3]);
+			if(isset($ra_kodeobat3)){
+				$jumracik3	= count($ra_kodeobat3);
+				for($ra3 = 0; $ra3 <= $jumracik3-1; $ra3++){
+					$nama_obat_ra3	= data_master("tbl_barang", array("kodebarang" => $ra_kodeobat3[$ra3]))->namabarang;
+					$harga_ra3		= str_replace(",", "", $ra_harga3[$ra3]);
+					$totalharga_ra3 = str_replace(",", "", $ra_totalharga3[$ra3]);
 
-				$data_racik3	= array(
-					"koders"		=> $cabang,
-					"orderno"		=> $orderno,
-					"racikid"		=> "RACIK3",
-					"noreg"			=> $noreg_dok,
-					"kodeobat"		=> $ra_kodeobat3[$ra3],
-					"namaobat"		=> $nama_obat_ra3,
-					"satuan"		=> $ra_satuan3[$ra3],
-					"qty_racik"		=> $ra_qtyracik3[$ra3],
-					"qty_jual"		=> $ra_qtyjual3[$ra3],
-					"harga"			=> $harga_ra3,
-					"totalharga"	=> $totalharga_ra3,
-					"aturanpakai"	=> $ra_aturanpakai3,
-					"proses"		=> 1,
-				);
+					$data_racik3	= array(
+						"koders"		=> $cabang,
+						"orderno"		=> $orderno,
+						"racikid"		=> "RACIK3",
+						"noreg"			=> $noreg_dok,
+						"kodeobat"		=> $ra_kodeobat3[$ra3],
+						"namaobat"		=> $nama_obat_ra3,
+						"satuan"		=> $ra_satuan3[$ra3],
+						"qty_racik"		=> $ra_qtyracik3[$ra3],
+						"qty_jual"		=> $ra_qtyjual3[$ra3],
+						"harga"			=> $harga_ra3,
+						"totalharga"	=> $totalharga_ra3,
+						"aturanpakai"	=> $ra_aturanpakai3,
+						"proses"		=> 1,
+					);
 
-				$query_insert_racik = $this->db->insert("tbl_eracik", $data_racik3);
+					$query_insert_racik = $this->db->insert("tbl_eracik", $data_racik3);
+				}
 			}
 
 			// -- ORDER PERIKSA -- //
@@ -1423,6 +1434,53 @@ class Poliklinik extends CI_Controller {
 
 	}
 
+	public function pcare_get_data_pas($noreg){
+		$trx		= $this->db->get_where("tbl_icdtr", ["noreg" => $noreg]);
+		$ins		= $this->db->get_Where("tbl_rekammedisrs", ["noreg" => $noreg]);
+
+		$res		= [];
+		$keluhan_awal = $tinggi_badan = $berat_badan = "";
+
+		switch(true){
+			case ($trx == false) : 
+				$status		= "error";
+				$message	= "Gagal memuat diagnosa";
+				break;
+			case ($ins == false) : 
+				$status		= "error";
+				$message	= "Gagal memuat data pasien";
+				break;
+			default : 
+				$data_pas	= $ins->row();
+
+				$status		= "success";
+				$message	= "";
+
+				$keluhan_awal	= $data_pas->keluhanawal;
+				$tinggi_badan	= $data_pas->tinggibadan;
+				$berat_badan	= str_replace(".00", "", $data_pas->beratbadan);
+
+				foreach($trx->result() as $t){
+					$res[]	= [
+						"icd_code"	=> $t->icdcode,
+						"icd_name"	=> data_master("tbl_icdinb", ["code" => $t->icdcode])->str
+					];
+				}
+				break;
+		}
+
+		echo json_encode([
+			"status"	=> $status,
+			"message"	=> $message,
+			"data_diag"	=> $res,
+			"data_pas"	=> [
+				"keluhan_awal"	=> $keluhan_awal,
+				"tinggi_badan"	=> $tinggi_badan,
+				"berat_badan"	=> $berat_badan
+			]
+		], JSON_UNESCAPED_SLASHES);
+	}
+
 	// View
 
 	public function pemeriksaan_perawat(){
@@ -1489,6 +1547,8 @@ class Poliklinik extends CI_Controller {
 				// $qcek2 = count($qcek);
 				// $ttv = $this->db->query("SELECT '-' as id ,'-' as koders ,'-' as noreg ,'-' as rekmed ,'-' as tglperiksa ,'-' as tglkeluar ,'-' as jam ,'-' as tujuan ,'-' as kodepos ,'-' as koderuang ,'-' as keluhanawal ,'-' as pfisik ,'-' as diagnosa ,'-' as simpul ,'-' as anjuran ,'-' as resep ,'-' as kodeicd ,'-' as kodeicd2 ,'-' as kodeicd3 ,'-' as keadaan_pulang ,'-' as ketpulang ,'-' as kodokter ,'-' as tglkonsul ,'-' as rcounter ,'0' as nadi ,'-' as nadi2 ,'-' as nafas ,'-' as tdarah ,'-' as tdarah1 ,'-' as suhu ,'-' as tinggibadan ,'-' as beratbadan ,'-' as bmi ,'-' as bmiresult ,'-' as tglkembali ,'-' as jamkembali ,'-' as nojanji ,'-' as dikonsulkepoli ,'-' as dikonsuldr ,'-' as jamdikonsul ,'-' as tgldikonsul ,'-' as noregkonsul ,'-' as surat1 ,'-' as surat2 ,'-' as surat3 ,'-' as surat4 ,'-' as alasanpulang ,'-' as urutkonsul ,'-' as oksigen ,'-' as kejiwaan ,'-' as lokalis ,'-' as rencana ,'-' as eye ,'-' as verbal ,'-' as movement ,'-' as nyeri ,'-' as urutin ,'-' as rmutama ,'-' as diagu ,'-' as diags ,'-' as tindu ,'-' as tinds ,'-' as asal ,'-' as gambar1 ,'-' as gambar2 ,'-' as pasienvk ,'-' as dead from tbl_rekammedisrs limit 1")->row();
 				$ttv = $this->db->query("SELECT * FROM tbl_rekammedisrs WHERE noreg = '$ceknoreg' and rekmed = '$rekmed' and koders='$koders'")->row();
+
+				$data_regist	= $this->db->query("SELECT * FROM pasien_rajal WHERE noreg = '$ceknoreg' AND rekmed = '$rekmed'");
 
 				// ICDTR //
 				$icd = $this->db->query("SELECT * FROM tbl_icdtr WHERE noreg = '$ceknoreg' and koders='$koders'")->num_rows();
@@ -1686,7 +1746,11 @@ class Poliklinik extends CI_Controller {
 					$data_riwayat_pasien	= $riwayat_pasien->row();
 				}
 
-				$aturan_pakai	= $this->db->query("SELECT * FROM tbl_barangsetup WHERE apogroup = 'ATURANPAKAI'");
+				// PCARE
+				$pcarepoli			= $this->db->get_where("bpjs_pcare_poli", ["poliSakit" => "1"]);
+				$pcaredr			= $this->db->get("bpjs_pcare_dokter");
+				$pcarestatuspulang	= $this->db->get("bpjs_pcare_status_pulang");
+				$pcarekesadaran     = $this->db->get("bpjs_pcare_kesadaran");
 
 				// VALUE
 				if(!empty($cek)){
@@ -1739,7 +1803,11 @@ class Poliklinik extends CI_Controller {
 						"total_erad"	=> $total_erad,
 						"status_kasir"	=> ($query_kasir == 0)? 0 : 1,
 						"riwayat_pasien"	=> $data_riwayat_pasien,
-						"aturan_pakai"	=> $aturan_pakai,
+						"data_regist"	=> $data_regist->row(),
+						"pcare_poli"	=> $pcarepoli,
+						"pcare_dr"		=> $pcaredr,
+						"pcare_sp"		=> $pcarestatuspulang,
+						"pcare_sadar"   => $pcarekesadaran,								
 					];
 					
 					$this->load->view("poliklinik/v_poliklinik_p_dokter_add", $data);
@@ -2274,4 +2342,5 @@ class Poliklinik extends CI_Controller {
 		echo $image;
 
 	}
+
 }
